backend/requirements.txt
chromadb==0.4.24
numpy<2
fastapi==0.110.0
openai>=1.40.0,<2
pandas==2.2.2
pyarrow==16.1.0
pydantic==2.7.0
pyyaml==6.0.1
uvicorn[standard]==0.29.0
